add redux folder, add redux types
