+++ components/hooks/useBreadcrumbs.tsx
@@ -1,9 +1,10 @@
 import { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 
-import { addBreadcrumbs, replaceBreadcrumbs } from '../../store/actions';
+import { addBreadcrumbs, replaceBreadcrumbs } from '../../store/redux/actions';
+import { InitialAppState } from '../../store/redux/types';
 
-import { InitialAppState, BreadcrumbsData } from '../../helpers/types';
+import { BreadcrumbsData } from '../../helpers/types';
 
 export const useBreadcrumbs = (
   currentBreadcrumbs: BreadcrumbsData,
