+++ pages/_app.tsx
@@ -6,10 +6,11 @@ import { Provider } from 'react-redux';
 import { ThemeProvider } from 'styled-components';
 import { throttle } from 'lodash';
 
-import appReducer from '@store/reducer';
+import appReducer from '@store/redux/reducer';
 
-import { InitialAppState } from '@helpers/types';
+import { InitialAppState } from '@store/redux/types';
 import { DEFAULT_BREADCRUMB } from '@store/statics';
+
 import '@helpers/Grid/Grid.scss';
 import { GlobalStyles } from '@helpers/GlobalStylesHelper';
 import { theme } from '@helpers/Theme';
